+++ graphistry/plotter.py
@@ -1,9 +1,10 @@
 from .PlotterBase import PlotterBase
-from .gremlin import GremlinMixin, CosmosMixin
+from .gremlin import GremlinMixin, CosmosMixin, NeptuneMixin
 
-class Plotter(CosmosMixin, GremlinMixin, PlotterBase, object):
+class Plotter(CosmosMixin, NeptuneMixin, GremlinMixin, PlotterBase, object):  # type: ignore
 
     def __init__(self, *args, **kwargs):
         PlotterBase.__init__(self, *args, **kwargs)
         GremlinMixin.__init__(self, *args, **kwargs)
         CosmosMixin.__init__(self, *args, **kwargs)
+        NeptuneMixin.__init__(self, *args, **kwargs)